Added map feature with user and station markers
Introduced Leaflet for map rendering in the Home component, dynamically displaying the user's location and the nearest weather station. This enhancement aims to provide users with a visual context for their current position relative to the chosen weather station, improving the app's usability and interactivity. Additionally, refined the presentation of the Home page with minor CSS adjustments to accommodate the new map view. SVG icons for the location dot and thumbtack have been added to symbolize the user and station positions, respectively, on the map. The added dependencies, `leaflet` and `react-leaflet`, along with necessary CSS and environmental setups, ensure seamless integration and responsive design.